+++ src/cards/fulcro/democards/manual_tests_of_dynamic_queries.cljs
@@ -20,13 +20,13 @@
   {:initial-state (fn [params] {:x 1 :y 42})
    :query         (fn [] [:x])
    :ident         (fn [] [:LEAF :ID])}
-  (dom/div nil
-    (dom/button #js {:onClick (fn [] (df/load this :leaf ui-leaf {:refresh [:root/leaf]}))} "Load a leaf using current query")
-    (dom/button #js {:onClick (fn [] (prim/set-query! this ui-leaf {:query [:x]}))} "Set query to :x")
-    (dom/button #js {:onClick (fn [] (prim/set-query! this ui-leaf {:query [:y]}))} "Set query to :y")
-    (dom/button #js {:onClick (fn [e] (if x
-                                        (m/set-value! this :x (inc x))
-                                        (m/set-value! this :y (inc y))))}
+  (dom/div
+    (dom/button {:onClick (fn [] (df/load this :leaf ui-leaf {:refresh [:root/leaf]}))} "Load a leaf using current query")
+    (dom/button {:onClick (fn [] (prim/set-query! this ui-leaf {:query [:x]}))} "Set query to :x")
+    (dom/button {:onClick (fn [] (prim/set-query! this ui-leaf {:query [:y]}))} "Set query to :y")
+    (dom/button {:onClick (fn [e] (if x
+                                    (m/set-value! this :x (inc x))
+                                    (m/set-value! this :y (inc y))))}
       (str "Count: " (or x y)))
     " Leaf"))
 
@@ -35,13 +35,14 @@
 (defsc Root [this {:keys [root/leaf] :as props}]
   {:initial-state (fn [p] {:root/leaf (prim/get-initial-state Leaf {})})
    :query         (fn [] [{:root/leaf (prim/get-query ui-leaf)}])}
-  (dom/div nil (ui-leaf leaf)))
+  (dom/div
+    (ui-leaf leaf)))
 
 (defcard-fulcro union-initial-app-state
   Root
   {}
   {:inspect-data true
-   :fulcro {:networking (server/new-server-emulator)}})
+   :fulcro       {:networking (server/new-server-emulator)}})
 
 (comment
   ; live manual test of query IDs and pulling query for live components
@@ -51,9 +52,7 @@
         component  (first (prim/ref->components reconciler [:LEAF :ID]))]
     [(prim/get-query-id component)
      (prim/get-query component @state)                      ;using component instance
-     (prim/get-query ui-leaf @state)                        ; using factory
-     ])
+     (prim/get-query ui-leaf @state)])                      ; using factory
 
-  (-> Leaf meta)
-  )
+  (-> Leaf meta))