+++ src/main/fulcro/client/localized_dom.clj
@@ -1,11 +1,11 @@
-(ns fulcro.client.alpha.localized-dom
+(ns fulcro.client.localized-dom
   (:refer-clojure :exclude [map meta time])
   (:require
     [clojure.future :refer :all]
     [fulcro.util :as util]
-    [fulcro.client.alpha.dom :as adom]
-    fulcro.client.alpha.dom-common
-    fulcro.client.alpha.localized-dom-common)
+    [fulcro.client.dom :as adom]
+    fulcro.client.dom-common
+    fulcro.client.localized-dom-common)
   (:import (cljs.tagged_literals JSValue)))
 
 (defn emit-tag [str-tag-name args]
@@ -13,38 +13,38 @@
         {attrs    :attrs
          children :children
          css      :css} conformed-args
-        css-props      (if css `(fulcro.client.alpha.localized-dom-common/add-kwprops-to-props nil ~css) nil)
+        css-props      (if css `(fulcro.client.localized-dom-common/add-kwprops-to-props nil ~css) nil)
         children       (mapv second children)
         attrs-type     (or (first attrs) :nil)              ; attrs omitted == nil
         attrs-value    (or (second attrs) {})]
     (case attrs-type
       :js-object
-      (let [attr-expr `(fulcro.client.alpha.localized-dom-common/add-kwprops-to-props ~attrs-value ~css)]
-        `(fulcro.client.alpha.dom/macro-create-element*
+      (let [attr-expr `(fulcro.client.localized-dom-common/add-kwprops-to-props ~attrs-value ~css)]
+        `(fulcro.client.dom/macro-create-element*
            ~(JSValue. (into [str-tag-name attr-expr] children))))
 
       :map
       (let [attr-expr (if (or css (contains? attrs-value :classes))
-                        `(fulcro.client.alpha.localized-dom-common/add-kwprops-to-props ~(adom/clj-map->js-object attrs-value) ~css)
+                        `(fulcro.client.localized-dom-common/add-kwprops-to-props ~(adom/clj-map->js-object attrs-value) ~css)
                         (adom/clj-map->js-object attrs-value))]
-        `(fulcro.client.alpha.dom/macro-create-element* ~(JSValue. (into [str-tag-name attr-expr] children))))
+        `(fulcro.client.dom/macro-create-element* ~(JSValue. (into [str-tag-name attr-expr] children))))
 
       :runtime-map
-      (let [attr-expr `(fulcro.client.alpha.localized-dom-common/add-kwprops-to-props ~(adom/clj-map->js-object attrs-value) ~css)]
-        `(fulcro.client.alpha.dom/macro-create-element*
+      (let [attr-expr `(fulcro.client.localized-dom-common/add-kwprops-to-props ~(adom/clj-map->js-object attrs-value) ~css)]
+        `(fulcro.client.dom/macro-create-element*
            ~(JSValue. (into [str-tag-name attr-expr] children))))
 
       (:symbol :expression)
-      `(fulcro.client.alpha.localized-dom/macro-create-element
+      `(fulcro.client.localized-dom/macro-create-element
          ~str-tag-name ~(into [attrs-value] children) ~css)
 
       ;; also used for MISSING props
       :nil
-      `(fulcro.client.alpha.dom/macro-create-element*
+      `(fulcro.client.dom/macro-create-element*
          ~(JSValue. (into [str-tag-name css-props] children)))
 
       ;; pure children
-      `(fulcro.client.alpha.localized-dom/macro-create-element
+      `(fulcro.client.localized-dom/macro-create-element
          ~str-tag-name ~(JSValue. (into [attrs-value] children)) ~css))))
 
-(adom/gen-dom-macros fulcro.client.alpha.localized-dom/emit-tag)
+(adom/gen-dom-macros fulcro.client.localized-dom/emit-tag)