+++ src/main/fulcro/client/localized_dom.cljs
@@ -1,9 +1,9 @@
-(ns fulcro.client.alpha.localized-dom
+(ns fulcro.client.localized-dom
   (:refer-clojure :exclude [map meta time])
   (:require
-    fulcro.client.alpha.dom
+    fulcro.client.dom
     fulcro.util
-    [fulcro.client.alpha.localized-dom-common :as cdom]))
+    [fulcro.client.localized-dom-common :as cdom]))
 
 (declare a abbr address area article aside audio b base bdi bdo big blockquote body br button canvas caption cite
   code col colgroup data datalist dd del details dfn dialog div dl dt em embed fieldset figcaption figure footer form
@@ -13,9 +13,9 @@
   td tfoot th thead time title tr track u ul var video wbr circle clipPath ellipse g line mask path
   pattern polyline rect svg text defs linearGradient polygon radialGradient stop tspan)
 
-(def node fulcro.client.alpha.dom/node)
-(def render-to-str fulcro.client.alpha.dom/render-to-str)
-(def create-element fulcro.client.alpha.dom/create-element)
+(def node fulcro.client.dom/node)
+(def render-to-str fulcro.client.dom/render-to-str)
+(def create-element fulcro.client.dom/create-element)
 
 (letfn [(arr-append* [arr x] (.push arr x) arr)
         (arr-append [arr tail] (reduce arr-append* arr tail))]
@@ -25,28 +25,28 @@
      (let [[head & tail] args]
        (cond
          (nil? head)
-         (fulcro.client.alpha.dom/macro-create-element*
+         (fulcro.client.dom/macro-create-element*
            (doto #js [type (cdom/add-kwprops-to-props #js {} csskw)]
              (arr-append tail)))
 
          (object? head)
-         (fulcro.client.alpha.dom/macro-create-element*
+         (fulcro.client.dom/macro-create-element*
            (doto #js [type (cdom/add-kwprops-to-props head csskw)]
              (arr-append tail)))
 
          (map? head)
-         (fulcro.client.alpha.dom/macro-create-element*
+         (fulcro.client.dom/macro-create-element*
            (doto #js [type (clj->js (cdom/add-kwprops-to-props head csskw))]
              (arr-append tail)))
 
-         (fulcro.client.alpha.dom/element? head)
-         (fulcro.client.alpha.dom/macro-create-element*
+         (fulcro.client.dom/element? head)
+         (fulcro.client.dom/macro-create-element*
            (doto #js [type (cdom/add-kwprops-to-props #js {} csskw)]
              (arr-append args)))
 
          :else
-         (fulcro.client.alpha.dom/macro-create-element*
+         (fulcro.client.dom/macro-create-element*
            (doto #js [type (cdom/add-kwprops-to-props #js {} csskw)]
              (arr-append args))))))))
 
-(fulcro.client.alpha.dom/gen-client-dom-fns fulcro.client.alpha.localized-dom/macro-create-element)
+(fulcro.client.dom/gen-client-dom-fns fulcro.client.localized-dom/macro-create-element)